Add initial draft of auth helpers. (#114)
* Add authentication helpers.

This adds helper functions for authenticating users. The helper
functions will extract a JWT from an HTTP header, decode and validate
the JWT, and perform a database lookup for a User object that matches
the JWT's subject.

* Add /users/current endpoint for getting info about currently logged in user.

* Add AUTH0_DOMAIN to docker-compose.yml.

This isn't a secret, and putting it here will make local development a
bit easier.